+++ examples/candle_whisper/Cargo.toml
@@ -10,6 +10,7 @@ version.workspace = true
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [features]
+default = []
 metal = ["candle-core/metal", "candle-nn/metal"]
 cuda = ["candle-core/cuda", "candle-nn/cuda", "candle-transformers/cuda", "dep:bindgen_cuda"]
 
@@ -28,8 +29,8 @@ anyhow = "1.0.79"
 
 candle-core = { version = "0.3.2", optional = true }
 candle-nn = { version = "0.3.2", optional = true }
-candle-transformers = { version = "0.3.2", optional = true }
-tokenizers = { version = "0.13.4", default-features = false }
+candle-transformers = { version = "0.3.2" }
+tokenizers = { version = "0.15.0", features = ["onig"] }
 rand = "0.8.5"
 serde_json = "1.0.109"
 byteorder = "1.5.0"
