+++ examples/candle_whisper/src/melfilters.bytes
