+++ examples/candle_whisper/src/melfilters128.bytes
