+++ examples/candle_whisper/src/multilingual.rs
@@ -1,4 +1,8 @@
+use crate::handlers::token_id;
+use crate::model::Model;
 use candle_core::{IndexOp, Result, Tensor, D};
+use candle_nn::ops::softmax;
+use candle_transformers::models::whisper::{self as m, audio, Config};
 use tokenizers::Tokenizer;
 
 const LANGUAGES: [(&str, &str); 99] = [
@@ -104,11 +108,7 @@ const LANGUAGES: [(&str, &str); 99] = [
 ];
 
 /// Returns the token id for the selected language.
-pub fn detect_language(
-    model: &mut super::Model,
-    tokenizer: &Tokenizer,
-    mel: &Tensor,
-) -> Result<u32> {
+pub fn detect_language(model: &mut Model, tokenizer: &Tokenizer, mel: &Tensor) -> Result<u32> {
     let (_bsize, _, seq_len) = mel.dims3()?;
     let mel = mel.narrow(
         2,
@@ -118,22 +118,22 @@ pub fn detect_language(
     let device = mel.device();
     let language_token_ids = LANGUAGES
         .iter()
-        .map(|(t, _)| crate::token_id(tokenizer, &format!("<|{t}|>")))
+        .map(|(t, _)| token_id(tokenizer, &format!("<|{t}|>")))
         .collect::<Result<Vec<_>>>()?;
-    let sot_token = crate::token_id(tokenizer, crate::m::SOT_TOKEN)?;
+    let sot_token = token_id(tokenizer, m::SOT_TOKEN)?;
     let audio_features = model.encoder_forward(&mel, true)?;
     let tokens = Tensor::new(&[[sot_token]], device)?;
     let language_token_ids = Tensor::new(language_token_ids.as_slice(), device)?;
     let ys = model.decoder_forward(&tokens, &audio_features, true)?;
     let logits = model.decoder_final_linear(&ys.i(..1)?)?.i(0)?.i(0)?;
     let logits = logits.index_select(&language_token_ids, 0)?;
-    let probs = candle_nn::ops::softmax(&logits, D::Minus1)?;
+    let probs = softmax(&logits, D::Minus1)?;
     let probs = probs.to_vec1::<f32>()?;
     let mut probs = LANGUAGES.iter().zip(probs.iter()).collect::<Vec<_>>();
     probs.sort_by(|(_, p1), (_, p2)| p2.total_cmp(p1));
     for ((_, language), p) in probs.iter().take(5) {
         println!("{language}: {p}")
     }
-    let language = crate::token_id(tokenizer, &format!("<|{}|>", probs[0].0 .0))?;
+    let language = token_id(tokenizer, &format!("<|{}|>", probs[0].0 .0))?;
     Ok(language)
 }
